+++ lib/Mvm/Allocator/gcmapper.h
@@ -12,6 +12,7 @@
 
 #include "gcerror.h"
 #include "osdep.h"
+#include "types.h"
 
 class GCMinAlloc;
 
@@ -19,28 +20,28 @@ class GCMin {
 public:
 	inline GCMin() {}
 	inline ~GCMin() {}
-	inline void *operator new(size_t);
-	inline void operator delete(void *, size_t);
+	inline void *operator new(uintptr_t);
+	inline void operator delete(void *, uintptr_t);
 };
 
 class GCMappedArea : public GCMin {
  	GCMappedArea	*_prev;
  	GCMappedArea	*_next;
  	void					*_mapped_area;
- 	size_t				_mapped_nbb;
+ 	uintptr_t				_mapped_nbb;
 
 public:
- 	static size_t round(size_t n) { return (n + PAGE_SIZE - 1) & -PAGE_SIZE; }
+ 	static uintptr_t round(uintptr_t n) { return (n + PAGE_SIZE - 1) & -PAGE_SIZE; }
 
  	GCMappedArea	*prev() { return _prev; }
  	GCMappedArea	*next() { return _next; }
 	
  	inline GCMappedArea() { _prev = _next = this; }
- 	inline GCMappedArea(size_t n) { _prev = _next = this; mmap(n); }
- 	inline GCMappedArea(GCMappedArea *p, size_t n) { append(p); mmap(n); }
+ 	inline GCMappedArea(uintptr_t n) { _prev = _next = this; mmap(n); }
+ 	inline GCMappedArea(GCMappedArea *p, uintptr_t n) { append(p); mmap(n); }
 
- 	inline void		initialise(void *p, size_t n) { _mapped_area = p; _mapped_nbb = n; }
- 	inline size_t nbb() const { return _mapped_nbb; }
+ 	inline void		initialise(void *p, uintptr_t n) { _mapped_area = p; _mapped_nbb = n; }
+ 	inline uintptr_t nbb() const { return _mapped_nbb; }
  	inline void		*area() { return _mapped_area; }
 
  	inline void append(GCMappedArea *p) {
@@ -56,23 +57,23 @@ public:
  	}
 	
  	GCMappedArea *	munmap();
- 	GCMappedArea *	mmap(size_t sz);
- 	int							mremap(size_t sz);
+ 	GCMappedArea *	mmap(uintptr_t sz);
+ 	int							mremap(uintptr_t sz);
 	
- 	static void *do_mmap(size_t sz);
- 	static void  do_munmap(void *, size_t sz);
+ 	static void *do_mmap(uintptr_t sz);
+ 	static void  do_munmap(void *, uintptr_t sz);
 };
 
 class GCMinAllocStack {
- 	unsigned int		current;
+ 	uintptr_t		current;
  	unsigned int		max;
  	unsigned int		*free_list;
  	unsigned short	nbb;
 	
 public:
- 	inline void initialise(size_t n) { nbb = n; current = max = 0; free_list = 0; }
+ 	inline void initialise(uintptr_t n) { nbb = n; current = max = 0; free_list = 0; }
 
-	inline void fill(unsigned int st, size_t n) {
+	inline void fill(unsigned int st, uintptr_t n) {
 		current = st;
 		max = st + n;
 	}
@@ -83,9 +84,9 @@ public:
  	}
 
  	inline void *alloc() {
- 		register unsigned int res;
+ 		register uintptr_t res;
  		if(free_list) {
- 			res = (unsigned int)free_list;
+ 			res = (uintptr_t)free_list;
  			free_list = *((unsigned int **)res);
  		} else {
  			res = current;
@@ -94,14 +95,14 @@ public:
  				return 0;
  		}
 		
- 		return (void *)res;
+ 		return (void*)res;
  	}
 };
 
 class GCMinAlloc {
- 	static const size_t	log_max_min_alloc = 6;											/* 64 octets */
- 	static const size_t	max_min_alloc = (1 << log_max_min_alloc);
- 	static const size_t log_step = 2;
+ 	static const uintptr_t	log_max_min_alloc = 6;											/* 64 octets */
+ 	static const uintptr_t	max_min_alloc = (1 << log_max_min_alloc);
+ 	static const uintptr_t log_step = 2;
 	
  	/* piles d'allocations */
  	static GCMinAllocStack			_stacks[max_min_alloc >> log_step];
@@ -116,11 +117,11 @@ public:
 	static void initialise();
 	static void destroy();
 	
- 	static inline void *minalloc(size_t);
- 	static inline void minfree(void *, size_t);
+ 	static inline void *minalloc(uintptr_t);
+ 	static inline void minfree(void *, uintptr_t);
 };
 
-inline void *GCMinAlloc::minalloc(size_t nbb) {
+inline void *GCMinAlloc::minalloc(uintptr_t nbb) {
  	GCMinAllocStack *s = stacks[nbb];
  	void *res = s->alloc();
 
@@ -129,15 +130,15 @@ inline void *GCMinAlloc::minalloc(size_t nbb) {
  	return res;
 }
 
-inline void GCMinAlloc::minfree(void *ptr, size_t nbb) {
+inline void GCMinAlloc::minfree(void *ptr, uintptr_t nbb) {
  	stacks[nbb]->inject_free(ptr);
 }
 
-inline void *GCMin::operator new(size_t nbb) { 
+inline void *GCMin::operator new(uintptr_t nbb) { 
  	return GCMinAlloc::minalloc(nbb);
 }
 
-inline void GCMin::operator delete(void *ptr, size_t nbb) { 
+inline void GCMin::operator delete(void *ptr, uintptr_t nbb) { 
 	GCMinAlloc::minfree(ptr, nbb);
 }
